for-each-ref: mark :remote-name and :remote-ref as experimental
These expansions will most likely be introduced into Git for Windows as
a test balloon first, and if they work out as expected, we'll try to get
the patches into upstream Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

squash! for-each-ref: mark :remote-name and :remote-ref as experimental

for-each-ref: mark :remotename and :remoteref as experimental

These expansions will most likely be introduced into Git for Windows as
a test balloon first, and if they work out as expected, we'll try to get
the patches into upstream Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>